+++ datar/dplyr/tidyselect.py
@@ -188,7 +188,8 @@ def matches(
 @register_func(context=Context.EVAL)
 def all_of(
         _data: DataFrame,
-        x: Iterable[Union[int, str]]
+        x: Iterable[Union[int, str]],
+        _base0: Optional[bool] = None
 ) -> List[str]:
     """For strict selection.
 
@@ -198,6 +199,8 @@ def all_of(
     Args:
         _data: The data piped in
         x: A set of variables to match the columns
+        _base0: Whether `x` is 0-based or not.
+            if not provided, will use `datar.base.getOption('index.base.0')`
 
     Returns:
         The matched column names
@@ -207,7 +210,9 @@ def all_of(
             in `_data` columns
     """
     all_columns = _data.columns
-    x = all_columns[vars_select(all_columns, x)]
+    x = all_columns[vars_select(all_columns, x, base0=_base0)]
+    # where do errors raise?
+
     # nonexists = setdiff(x, all_columns)
     # if nonexists:
     #     raise ColumnNotExistingError(
@@ -222,7 +227,8 @@ def any_of(
         _data: DataFrame,
         x: Iterable[Union[int, str]],
         # pylint: disable=redefined-builtin
-        vars: Optional[Iterable[str]] = None
+        vars: Optional[Iterable[str]] = None,
+        _base0: Optional[bool] = None
 ) -> List[str]:
     """Select but doesn't check for missing variables.
 
@@ -232,19 +238,22 @@ def any_of(
     Args:
         _data: The data piped in
         x: A set of variables to match the columns
+        _base0: Whether `x` is 0-based or not.
+            if not provided, will use `datar.base.getOption('index.base.0')`
 
     Returns:
         The matched column names
     """
     vars = vars or _data.columns
-    x = vars_select(vars, x, raise_nonexists=False)
-    exists = []
-    for idx in x:
-        try:
-            exists.append(vars[idx])
-        except IndexError:
-            ...
-    return intersect(vars, exists)
+    x = vars_select(vars, x, raise_nonexists=False, base0=_base0)
+    # exists = []
+    # for idx in x:
+    #     try:
+    #         exists.append(vars[idx])
+    #     except IndexError:
+    #         ...
+    # do we need intersect?
+    return intersect(vars, [vars[idx] for idx in x])
 
 @register_func(None)
 def num_range(
@@ -304,4 +313,5 @@ def _filter_columns(
                     column.lower() if ignore_case else column
             )):
                 ret.append(column)
+
     return ret