+++ tests/test_dplyr_bind.py
@@ -1,7 +1,7 @@
 # https://github.com/tidyverse/dplyr/blob/master/tests/testthat/test-bind.R
-from pipda.utils import Expression
 import pytest
 
+from pandas.testing import assert_frame_equal
 from datar.all import *
 
 def test_handle_dict():
@@ -45,12 +45,12 @@ def test_bind_col_null():
 
 def test_repair_names():
     df = tibble(a = 1, b = 2)
-    bound = bind_cols(df, df)
+    bound = bind_cols(df, df, _base0=True)
     assert bound.columns.tolist() == ['a__0', 'b__1', 'a__2', 'b__3']
 
     t1 = tibble(a=1)
     t2 = tibble(a=2)
-    bound = bind_cols(t1, t2)
+    bound = bind_cols(t1, t2, _base0=True)
     assert bound.columns.tolist() == ['a__0', 'a__1']
 
 def test_incompatible_size_fill_with_NA():
@@ -139,6 +139,9 @@ def test_bind_factors():
     assert out.a.cat.categories.tolist() == ["a"]
     assert out.a.astype(object).fillna("NA").tolist() == ["a", "NA"]
 
+    out2 = bind_rows(None, [df1, df2])
+    assert_frame_equal(out2, out)
+
 def test_bind_na_cols():
     df1 = tibble(x=factor(["foo", "bar"]))
     df2 = tibble(x=NA)
@@ -174,9 +177,9 @@ def test_create_id_col():
     df1 = df >> head(3)
     df2 = df >> tail(2)
     out = df1 >> bind_rows(df2, _id='col')
-    assert out.col.tolist() == [0,0,0,1,1]
+    assert out.col.tolist() == [1,1,1,2,2]
 
-    out = bind_rows([df1, df2], _id='col')
+    out = bind_rows([df1, df2], _id='col', _base0=True)
     assert out.col.tolist() == [0,0,0,1,1]
 
     out = bind_rows(None, one=df1, two=df2, _id="col")